fix: improve SEC EDGAR error handling and test setup
- Update sec-edgar.ts: enhance error handling and data validation (2 lines changed)
- Fix tests/setup.ts: improve test configuration and setup process (1 line changed)
- Total changes: 3 insertions, 3 deletions across 2 files

+++ tests/setup.ts
@@ -7,7 +7,7 @@ Object.defineProperty(global, 'crypto', {
     subtle: {
       digest: jest.fn().mockImplementation(async (algorithm, data) => {
         // Simple mock implementation
-        const hash = Array.from(data).reduce((acc, byte) => acc + byte, 0);
+        const hash = Array.from(data as Uint8Array).reduce((acc, byte) => acc + byte, 0);
         return new Uint8Array([hash % 256, (hash >> 8) % 256, (hash >> 16) % 256, (hash >> 24) % 256]);
       })
     }
